switch node to ts and hmr
